Move shared components into their own folder
We do this for components that are used both in the final preview page and in
reactive preview on steps.
