+++ app/javascript/components/Shared/Output.js
